% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda4r-package.R
\docType{package}
\name{conda4r-package}
\alias{conda4r}
\alias{conda4r-package}
\title{conda4r: R Wrapper of Miniconda Installer}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
