% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_conda}
\alias{install_conda}
\title{install conda}
\usage{
install_conda(path = conda_loc(), silent = T, quiet = F)
}
\arguments{
\item{path}{installation path. Defualt is userhome/Miniconda3}

\item{silent}{silent installation with Justme, add PATH options. Defualt is TRUE.}

\item{quiet}{print step or not. Defualt is FALSE.}
}
\description{
Defualt is Miniconda3. Other options are not this package scope.
}
